+++ src/Symfony/Component/Notifier/Bridge/Firebase/Notification/WebNotification.php
@@ -12,17 +12,23 @@
 namespace Symfony\Component\Notifier\Bridge\Firebase\Notification;
 
 use Symfony\Component\Notifier\Bridge\Firebase\FirebaseOptions;
+use Symfony\Component\Notifier\Bridge\Firebase\TargetType;
 
 final class WebNotification extends FirebaseOptions
 {
+    public function __construct(string $target, array $options = [], array $data = [], TargetType $targetType = TargetType::Topic)
+    {
+        trigger_deprecation('symfony/firebase-notifier', '8.2', 'Using %s class is deprecated, use %s instead.', self::class, FirebaseOptions::class);
+
+        parent::__construct($target, ['notification' => $options], $data, $targetType);
+    }
+
     /**
      * @return $this
      */
     public function icon(string $icon): static
     {
-        $this->options['icon'] = $icon;
-
-        return $this;
+        return $this->addWebpushOption('icon', $icon);
     }
 
     /**
@@ -30,8 +36,6 @@ final class WebNotification extends FirebaseOptions
      */
     public function clickAction(string $clickAction): static
     {
-        $this->options['click_action'] = $clickAction;
-
-        return $this;
+        return $this->addWebpushOption('click_action', $clickAction);
     }
 }