+++ src/Symfony/Component/Notifier/Bridge/Firebase/CHANGELOG.md
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+8.2
+---
+
+ * Send messages through the Firebase Cloud Messaging v1 API
+ * Deprecate `AndroidNotification`, `IOSNotification` and `WebNotification`, use `FirebaseOptions` instead
+ * Deprecate the `firebase://USERNAME:PASSWORD@default` DSN, use `firebase://PROJECT_ID?client_email=...&private_key_id=...&private_key=...` instead
+ * Deprecate the `$token` argument of `FirebaseTransport::__construct()`
+
 5.3
 ---